lazy load createpackage component
